pyproject.toml
[tool.poetry]
name = "djangoplatformforstady"
version = "0.1.0"
description = ""
authors = ["Fish3r228 <zayarobert30@gmail.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.5"
djangorestframework = "^3.16.1"
pillow = "^11.3.0"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
stripe = "^12.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
